Added enemy prefab and tags, minor bugfix to respawn script
Created the enemy prefab, the "Enemy" gameobject should always be at 0,0,0. To move the enemy to a specific point, move the EnemyRespawnPoint object to the desired location. Also, a quick preventative bugfix to the Respawn script, added an if statement that ensures that the respawn only occurs when it's a Player that enters the collider.