Implemented level reload on respawn and falling/disappearing platform.
Made it so that the level reloads in the respawn script instead of just sending the player to a respawn point. Also added a functional falling/disappearing platform. The platform still needs some aesthetic work, however, as it currently just disappears 1 second after the player touches it with no indication that it will do so.