Made it so that the enemy now uses the proper sprite
The enemy now uses the skeleton sprite. Some changes had to be made to the fov script to allow for the origin of the fov to be moved.  A new Serialized Field was added that allows for changes in the x and y position of the origin of the fov.