Added dash mechanic which works horizontally, vertically, and diagonally
Works by setting the characters velocity to a higher value than its maximum velocity in the direction which the player is inputting. Dash using left shift. The player can dash every quarter of a second so long as they've touched the ground since the last time they dashed.

The vertical dashes feel a little weird at the moment, definitely need some polishing. Considering increasing gravity on upward dashes so that it doesn't feel as floaty, which will also allow for the upward dash to have a higher initial velocity.