Small update to comments and jump code
Theoretically, the code I added should prevent the character from double jumping. But that's not what's happening. I've tried serializing the isJumping variable to see why it's letting me double jump sometimes, but whenever I have the value of isJumping variable on my screen, I can never double jump. However, the moment the isJumping variable is off of my screen, mind you I still have the inspector open for the character, I just didn't scroll down to be able to see the value of isJumping, I can double jump again. I'm struggling to come up with a reason for this, my only thought is that maybe it has to do with the frame rate, and that when I have the inspector open and am looking at isJumping my framerate drops, meaning that when the character leaves the ground, they're further from the ground by the next frame. But I don't really think that's the reason...