Bat Movement upgrades
Bat now moves in a sin pattern rather than a striaght line. Options to change sin wave's height, speed, amplitude are all changeable. Also an variable named "oneForHorz", which has a field that changes if the wave is horizontal or vertical depending on if you have a one in there. It defaults to vertical.