Fixed multiple issues with character movement and started working on new level
Made it so that the bottom of the character uses a box collider instead of a circle collider, this prevents the player from getting stuck on corners. Then made sure that whenever the spacebar was not pressed isJumping was false. Made sure that holding the spacebar while sprinting wouldn't allow the player to jump, dash, then continue jumping after the dash finished. Finally, made a new level and began adding obstacles