Ethan Level 2 (Missing background)
My 2nd level.  I am having some issues with the background tilegrid, so I will need to ask someone at our meeting tonight (12/2/21) about it. Otherwise, it is good to go and implement into the level progression.

This level has two parts. The first part is a jumping part where the player must rely on their dash ability to progress successfully. If they fall, there is a crystal at the bottom to allow them to reset. But beware! There are spikes in the pit that the player needs to think quickly to avoid.

The 2nd part is a bat cave. After a short fall, the level will fade into the newer tile pallet, which is darker. This is to emulate the darkness of the deeper part of the cave. It is straight forward and easy to navigate. There are a few jumps, and 3 total bats the player must avoid to reach the end. That last bat in particular is tricky!